Moved Bootstrap to bootstrap.py
