+++ pythonforandroid/util.py
@@ -104,3 +104,24 @@ class JsonStore(object):
         else:
             with io.open(self.filename, 'w', encoding='utf-8') as fd:
                 fd.write(unicode(json.dumps(self.data, ensure_ascii=False)))
+
+
+def which(program, path_env):
+    '''Locate an executable in the system.'''
+    import os
+
+    def is_exe(fpath):
+        return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
+
+    fpath, fname = os.path.split(program)
+    if fpath:
+        if is_exe(program):
+            return program
+    else:
+        for path in path_env.split(os.pathsep):
+            path = path.strip('"')
+            exe_file = os.path.join(path, program)
+            if is_exe(exe_file):
+                return exe_file
+
+    return None
